user: Accept walrus+SHARD as full client
Together, walrus and SHARD implement all the functionality we need
from siad -- *except* one little convenience: resolving truncated
host keys, e.g. f1e23ab7. To do this, we need access to the full list
of hosts, but SHARD only lets us lookup a host that we already know
the full public key for.

For now, I think it's okay to just return an error if you try to use
walrus+SHARD to resolve a pubkey, but honestly I should probably just
add this functionality to SHARD. Prefix lookups can be done with
either a sorted list + binary search, or a radix tree; the latter is
more complicated, but avoids the cost of keeping the list sorted.
I'll probably go with the sort+search, since it's much easier to
verify correctness and the performance difference is probably small.