renter: Never extract MetaFiles to disk
I originally designed the metafile code to handle very large files,
large enough that they couldn't be kept in memory. This was probably
a case of premature optimization. A metafile is about 2000x smaller
than the file it references, so storing them in memory won't be a
problem until you go above ~10 TB. I can revisit this code when
someone actually needs to store a 10 TB file and can't afford 64 GB
of RAM.

Anyway, this also vastly simplifies the renter and filesystem code,
leaves less garbage on disk, and makes the atomicity more obvious.
The only real downside is that if the process dies before you save,
the new metafile isn't on disk at all, whereas before, there was a
(negligible) chance that you'd be able to recover it. Also, modifying
the shard slices means we need to work with a *[]SectorSlice (e.g. in
ShardUploader), which is a little gross.