renter: Increase max contract size to 2048
To my great annoyance, contracts do not fit in 1024 bytes. Perhaps
they could, if we trimmed some of the fat off of the Revision and
Signatures, but it's probably best to preserve the Sia-compatibility
of those objects.

The good news is that only the revision data would be lost; since the
header is intact, any affected contracts can be recovered.

This almost certainly breaks compatibility, but you can fix it with a
simple `truncate -s 2048`, so I'm not inclined to bother changing the
version number. Maybe someday us will have enough users that I'll
have to be more responsible with such things, but that day is not
today.