renterutil: (Re-)connect to hosts lazily
If this works as I hope, it will kill two birds with one stone
(err -- *feed* two birds with one *scone* -- sorry Jocelyn).
First, since (*HostSet).acquire is called in a goroutine, we should
connect to all hosts in parallel. Second, we should now be immune
to the problem of hosts closing the RPC loop early, since we close
and reopen the loop each time acquire is called. This is inefficient,
of course, but should be tolerable; reads and writes aren't expected
to be super-low-latency anyway.