wallet: Remove ChainStore from Store
I think this design is *slightly* better, but still needs more work.
I like separating subscription from the rest of the wallet behavior;
imagine a meta-wallet that manages multiple SeedWallets. Previously,
each SeedWallet would need to subscribe to consensus, whereas in the
new design, only the meta-wallet needs to subscribe.

Still, the lockedChainScanner business is not pretty. The fundamental
problem is that the ChainScanner needs to call OwnsAddress on the
wallet multiple times in the course of processing a block, so if
the wallet is concurrently adding/removing addresses while that's
happening, you can end up with a "partially-processed" block, e.g.
a block where one of the transactions was marked as relevant, but
an earlier transaction referencing the same address was not.

Technically, this shouldn't happen in practice, since you're not
supposed to add addresses that have already been used, and
removing an address requires a rescan. But that's not quite airtight
enough for me. So we're left with the lockedChainScanner hack, which
borrows the wallet's mutex to ensure that calls to OwnsAddress are
consistent throughput the processing of a single block.