wallet: Use SeedAddressInfo instead of []byte
The watch-only wallet was originally intended to be maximally
general, but I now consider this a premature optimization. You
almost never want to associate something other than this info
with an address, and by mandating it, we can make useful assumptions
elsewhere (such as: "the wallet will always know the unlock
conditions for every address it's tracking").