wallet: reify WatchOnlyWallet
This is a pretty radical and messy change. The WatchOnlyWallet is
now the "primary" wallet exported by this package, while the old
SeedWallet has been almost completely removed: it lives on as a
small HotWallet wrapper around SeedWallet. (Honestly, the wrapper
is so small that I may remove it entirely.) The idea is that a hot
wallet is just a seed plus a watch-only wallet, so reify the watch-
only wallet. As a consequence of this, the SeedManager has become
fairly useless; it is mainly useful for tracking the current seed
index and looking up the SeedAddressInfo for a particular address,
and SeedWallet can do both of those just fine. Something like the
SeedManager will be useful for things like load-from-seed, but I
figure I'll just end up rewriting that functionality as the need
arises.

Overall, I'm hoping that this refactor simplifies the wallet API
while also pushing people towards a more secure default (i.e. away
from hot wallets).