proto: Add 5% price leeway in Write
I've never been able to get to the bottom of why this happens. If
we're using the most recent host settings, underpayment should only
happen if the renter and host are many blocks out of sync. Oh well --
it clearly *is* still happening, so this band-aid is necessary until
the root cause can be determined.