renter: Remove ContractEditor
Going forward, contracts will be treated more "ephemerally" -- a
Session will no longer automatically call SetRevision (although
the latest revision can be accessed via a new Revision method).
This makes it easier to build contract servers, and to transition
to payment codes later on. It might seem like we lose some security
by not saving the revisions, but in reality we *exclusively* trust
the host's reported revision anyway, since the host can't benefit
from lying. This might change later (if a host can refund a renter,
for example), but we'll cross that bridge when we come to it.