proto: Precompute sector roots in Write
This makes good use of the CPU time that we would otherwise spend
waiting around for hosts to reply. More importantly, this design
means we only calculate the root once, instead of in both Write
and SectorBuilder.Finish. Benchmarks show significant improvement:
80 MB/s -> 120 MB/s in BenchmarkFileSystemWrite.

Calculating sector roots seems to be the slowest operation in the
pipeline, at least in terms of CPU time. That's a bummer. If we
had a hash function with a 64-byte block size and a special tree
hashing flag, we could probably speed up the root calculation by
2x or more. Unfortunately, no such hash function exists (that I'm
aware of, so we're stuck with BLAKE2b for the foreseeable future.