renter: Add BenchmarkWriteMetaFile
On my MacBook's SSD, we can write 250,000 SectorSlices in just
over 1 second. That speed seems acceptable to me (not many people
are storing multiple TB on Sia yet), but down the road we'll
definitely need to optimize a bit. The biggest bottleneck here is
the fact that we rewrite the entire metafile each time we upload
a sector, so you end up doing O(n^2) I/O. Switching to an append-
only algorithm should greatly improve performance, but will require
careful handling to preserve atomicity.