proto: Return contract txn instead of broadcasting
On more than one occasion, I've seen the tpool reject transaction
sets that the host accepted. Given that the host (presumably)
submitted the set to its own tpool without error, we can assume
that the host will honor the contract, so being rejected by our
own tpool should result in a warning at most. Consequently, I'm
punting the responsibility for handling this to the caller.

This means that, in theory, we could replace the tpool parameter
with a simple feePerByte parameter (since computing the fee is the
only thing we use the tpool for now). I'm going to leave it for
now, but I'll make a note to reevaluate this in 6-12 months.