renterhost: Enable raw message streaming
This is some gnarly code, and it took quite some time to get it
right, but ultimately I'm quite happy with it. Chaining all those
Readers and Writers together is really satisfying.

Anyway, this lets callers stream RPC messages without authenticating
them first. Which is generally a bad idea -- there's a reason AEAD is
a thing! But for our purposes, this allows us to start hashing sector
data immediately, which is a Big Deal for performance (since BLAKE2b
is currently our primary bottleneck). You just need to be careful to
not do anything reckless with the data before verifying it, like
writing it directly to disk or streaming it to a client over HTTP.
Fortunately, Sia's design makes this difficult to do anyway; we
still wait for the full message to arrive before proceeding to the
erasure decoding step, so as long as verify the poly1305 tag before
decoding, there's no loss of security.

I do still feel a bit guilty about introducing API surface that, if
misused, could constitute an attack vector. If our hashing was on
par with our encryption and erasure coding (that is, >10 Gbps), I
would happily skip this optimization. Alas.