reedsolomon: Remove inversion tree
The purpose of the inversion tree is to cache matrices. This saves
time if you're reconstructing many matrices with the same pattern
of missing shards. But the amount of time saved is quite small
compared to the amount of code required (benchmarks show no change,
maybe even 1% faster without the inversion tree), so I'm deleting
it for simplicity.