ed25519: Reduce package scope, rename to ed25519hash
Since crypto/ed25519 is now in the standard library, it seems wrong
to be duplicating most of it, especially types like PrivateKey and
PublicKey. Instead, make our package operate on stdlib types, keeping
the optimized functions for signing and verifying hashes.

This makes the API slightly less nice; we can't provide SignHash and
VerifyHash methods on the key types, because we aren't defining those
types anymore. The only thing that really bugs me about this is that
the stdlib (ed25519.PrivateKey).PublicKey method does *not* return an
ed25519.PublicKey -- it returns a crypto.PublicKey, which wraps the
former in an interface{} indirection. y.

Consequently, I went ahead and removed the HashSigner/HashVerifier
interfaces in renterhost, since they weren't all that useful anyway;
renterhost now operates on ed25519 keys explicitly. I also removed
(hostdb.HostPublicKey).VerifyHash, since it was only being used in
one or two places and it's trivial to do Verify(hpk.Ed25519()) at
the call site.

I'm pretty happy with the new design, because it makes ed25519hash
a purely optional optimization. If you don't know it exists, you can
just use the stdlib ed25519 functions to sign and verify -- and
later, drop it in for a nice 2x performance boost.