proto: Verify host's revision signatures
Technically there's no real incentive for the renter to do this;
as usual, if the host screwed up, it just means the renter got to
call an RPC for free. Besides, we can't even tell the host that
their signature is invalid. But since we're already going to the
trouble of storing the signatures, and since SubmitContractRevision
exists, and since signature verification is pretty fast, we might
as well do it.